InstCombine: remove one of the RetDec modifications.
This modification was inhibiting the following store instruction
optimizations:
  %tmp2 = bitcast float %tmp1 to i32
  store i32 %tmp2, i32* @g
==>
  store float %tmp1, float* bitcast (i32* @g to float*)

We still don't want to do this optimization, but disabling it caused
-instcombine to cycle infinitely in the new LLVM 8.0.0. Rather than
hacking LLVM some more in order to fix this, we decided to allow the
optimization (i.e. remove RetDec inhibition code) and revert its effects
afterwards in our own pass.